First pass at converting vote scraper
