+++ billy_metadata/mn.py
@@ -50,13 +50,11 @@ metadata = {
             'type':'primary',
             'display_name': '2009-2010 Regular Session',
             '_scraped_name': '86th Legislature, 2009-2010',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls86.asp',
         },
         '2010 1st Special Session': {
             'type': 'special',
             'display_name': '2010, 1st Special Session',
             '_scraped_name': '86th Legislature, 2010 1st Special Session',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls8620101.asp',
         },
         '2010 2nd Special Session': {
             'type':'special',
@@ -67,49 +65,41 @@ metadata = {
             'type':'primary',
             'display_name': '2011-2012 Regular Session',
             '_scraped_name': '87th Legislature, 2011-2012',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls87.asp',
         },
         '2011s1': {
             'type': 'special',
             'display_name': '2011, 1st Special Session',
             '_scraped_name': '87th Legislature, 2011 1st Special Session',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls8720111.asp',
         },
         '2012s1': {
             'type': 'special',
             'display_name': '2012, 1st Special Session',
             '_scraped_name': '87th Legislature, 2012 1st Special Session',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls8720121.asp',
         },
         '2013-2014': {
             'type': 'primary',
             'display_name': '2013-2014 Regular Session',
             '_scraped_name': '88th Legislature, 2013-2014',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls88.asp',
         },
         '2013s1': {
             'type': 'special',
             'display_name': '2013, 1st Special Session',
             '_scraped_name': '88th Legislature, 2013 1st Special Session',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls8820131.asp',
         },
         '2015-2016': {
             'type': 'primary',
             'display_name': '2015-2016 Regular Session',
             '_scraped_name': '89th Legislature, 2015-2016',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls89.asp',
         },
         '2015s1': {
             'type': 'special',
             'display_name': '2015, 1st Special Session',
             '_scraped_name': '89th Legislature, 2015 1st Special Session',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls8920151.asp',
         },
         '2017-2018': {
             'type': 'primary',
             'display_name': '2017-2018 Regular Session',
             '_scraped_name': '90th Legislature, 2017-2018',
-            'votes_url': 'http://www.house.leg.state.mn.us/votes/getVotesls90.asp',
         },
     },
     '_ignored_scraped_sessions': [
